+++ app/root.tsx
@@ -22,6 +22,9 @@ function TopNav() {
       <NavLink to="/explore" className={linkClass}>
         Explore
       </NavLink>
+      <NavLink to="/work" className={linkClass}>
+        Work
+      </NavLink>
     </nav>
   );
 }